+++ src/main/java/shop/catchmind/auth/application/AuthService.java
@@ -1,31 +1,11 @@
 package shop.catchmind.auth.application;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import shop.catchmind.auth.dto.AuthenticationDto;
-import shop.catchmind.member.domain.Member;
-import shop.catchmind.member.repository.MemberRepository;
+import shop.catchmind.auth.dto.request.IsExistedEmailRequest;
+import shop.catchmind.auth.dto.request.SignUpRequest;
+import shop.catchmind.auth.dto.response.IsExistedEmailResponse;
 
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class AuthService implements UserDetailsService {
+public interface AuthService {
+    void signUp(SignUpRequest request);
 
-    private final MemberRepository memberRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(final String email) throws UsernameNotFoundException {
-        Member member = memberRepository.findByEmail(email)
-                .orElseThrow(RuntimeException::new);// TODO: Exception 처리 필요
-
-        return AuthenticationDto.builder()
-                .id(member.getId())
-                .email(member.getEmail())
-                .password(member.getPassword())
-                .build();
-    }
+    IsExistedEmailResponse isExistedEmail(IsExistedEmailRequest request);
 }
