+++ src/main/java/shop/catchmind/auth/application/LoginService.java
@@ -1,0 +1,6 @@
+package shop.catchmind.auth.application;
+
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+public interface LoginService extends UserDetailsService {
+}
