+++ src/main/java/shop/catchmind/auth/application/LoginServiceImpl.java
@@ -1,0 +1,30 @@
+package shop.catchmind.auth.application;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import shop.catchmind.auth.dto.AuthenticationDto;
+import shop.catchmind.member.domain.Member;
+import shop.catchmind.member.repository.MemberRepository;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class LoginServiceImpl implements LoginService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(final String email) throws UsernameNotFoundException {
+        Member member = memberRepository.findByEmail(email)
+                .orElseThrow(RuntimeException::new);// TODO: Exception 처리 필요
+
+        return AuthenticationDto.builder()
+                .id(member.getId())
+                .email(member.getEmail())
+                .password(member.getPassword())
+                .build();
+    }
+}
